experimenting with relocatable branch
To experiment with Dirk Thomas' commits to the various ROS
repositories, this commit updates recipes catkin, gencpp,
genlisp, genmsg, genpy, ros and ros-comm to the relocatable
branch of the respective repository.
Furthermore, it removes patches and their reference from recipes
as they should not be required anymore with the new commits.

This commit still raises some issues to be resolved and is thought
as a base for discussion and continuous work on this topic.
It should not be merged into the main repository at this point.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@oss.bmw-carit.de>